sizeof(struct bcode) : 48 -> 40 bytes
    PUBLISHED_FROM=03bf338101fe5c788a84f3bdc8a31b794e9b43ab

+++ v7/v7.c
@@ -3762,8 +3762,8 @@ struct bcode {
   struct mbuf ops;   /* instruction opcode */
   struct mbuf lit;   /* literal table */
   struct mbuf names; /* function name, `args` arg names, local names */
-  int refcnt;        /* reference count */
-  int args;          /* number of args */
+  uint8_t refcnt;    /* reference count */
+  uint8_t args;      /* number of args */
 
   unsigned int strict_mode : 1;
   /*
